fix: moved project to treasury folder, exported the governance token as ownTokens
