--- projects/createDAO/config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  ethereum: {
-    factoryAddress: '0xc852E5Cb44C50614a82050163aB7170cB88EB5F9',
-    fromBlock: 24424853,
-  },
-};
